CAMUS_same/CUReg/requirements.txt
absl-py==0.15.0
addict==2.4.0
aiohttp @ file:///croot/aiohttp_1701112538292/work
aiosignal @ file:///tmp/build/80754af9/aiosignal_1637843061372/work
apptools==5.2.1
astor==0.8.1
asttokens==2.4.1
astunparse==1.6.3
async-timeout @ file:///opt/conda/conda-bld/async-timeout_1664876359750/work
attrs @ file:///croot/attrs_1695717823297/work
backcall==0.2.0
blinker @ file:///croot/blinker_1696539051175/work
Brotli @ file:///tmp/abs_ecyw11_7ze/croots/recipe/brotli-split_1659616059936/work
cachetools @ file:///tmp/build/80754af9/cachetools_1619597386817/work
certifi @ file:///home/conda/feedstock_root/build_artifacts/certifi_1707022139797/work/certifi
cffi @ file:///tmp/build/80754af9/cffi_1636541934635/work
chamfer-3D==0.0.0
chardet==5.2.0
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click @ file:///croot/click_1698129812380/work
cloudpickle @ file:///home/conda/feedstock_root/build_artifacts/cloudpickle_1697464713350/work
coloredlogs==15.0.1
comm==0.2.2
ConfigArgParse==1.7
configobj==5.0.8
contourpy==1.1.1
cryptography @ file:///croot/cryptography_1694444244250/work
cycler==0.12.1
cytoolz @ file:///home/conda/feedstock_root/build_artifacts/cytoolz_1706897079399/work
dash==2.17.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dask @ file:///croot/dask-core_1683065217061/work
decorator==5.1.1
easydict==1.13
emd-cuda==0.0.0
envisage==7.0.3
executing==2.0.1
fastjsonschema==2.19.1
Flask==3.0.3
flatbuffers==1.12
fonttools==4.47.0
frozenlist @ file:///croot/frozenlist_1698702560391/work
fsspec @ file:///home/conda/feedstock_root/build_artifacts/fsspec_1707102468451/work
fvcore==0.1.5.post20221221
gast==0.3.3
gendr==0.1.0
geomloss==0.2.6
-e git+https://github.com/qinzheng93/GeoTransformer.git@e7a135af4c318ff3b8d7f6c963df094d7e4ea540#egg=geotransformer
gitdb==4.0.11
GitPython==3.1.43
google-auth @ file:///croot/google-auth_1694152708165/work
google-auth-oauthlib==0.4.6
google-pasta @ file:///Users/ktietz/demo/mc3/conda-bld/google-pasta_1630577991354/work
grpcio==1.32.0
h5py @ file:///tmp/build/80754af9/h5py_1593454122442/work
humanfriendly==10.0
idna @ file:///croot/idna_1666125576474/work
imagecodecs-lite @ file:///home/conda/feedstock_root/build_artifacts/imagecodecs-lite_1695649144722/work
imageio==2.33.1
importlib-metadata @ file:///croot/importlib-metadata_1678997070253/work
importlib-resources==6.1.1
iopath==0.1.10
ipython==8.12.3
ipywidgets==8.1.2
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter_core==5.7.2
jupyterlab_widgets==3.0.10
Keras-Preprocessing @ file:///tmp/build/80754af9/keras-preprocessing_1612283640596/work
kiwisolver==1.4.5
knn-pytorch==0.1
lazy_loader==0.3
locket @ file:///home/conda/feedstock_root/build_artifacts/locket_1650660393415/work
Markdown @ file:///croot/markdown_1671541909495/work
MarkupSafe @ file:///opt/conda/conda-bld/markupsafe_1654597864307/work
matplotlib==3.7.4
matplotlib-inline==0.1.7
mayavi==4.8.1
mkl-fft==1.3.1
mkl-random @ file:///tmp/build/80754af9/mkl_random_1626186064646/work
mkl-service==2.4.0
multidict @ file:///croot/multidict_1701096859099/work
nbformat==5.10.4
nest-asyncio==1.6.0
networkx @ file:///home/conda/feedstock_root/build_artifacts/networkx_1680692919326/work
nibabel==5.2.1
numpy==1.24.4
numpy-stl==3.1.1
oauthlib @ file:///croot/oauthlib_1679489621486/work
open3d==0.18.0
opencv-python==4.8.1.78
OpenEXR @ file:///home/conda/feedstock_root/build_artifacts/openexr-python_1695649565706/work
opt-einsum @ file:///tmp/build/80754af9/opt_einsum_1621500238896/work
packaging @ file:///croot/packaging_1693575174725/work
pandas==2.0.3
parso==0.8.4
partd @ file:///home/conda/feedstock_root/build_artifacts/partd_1695667515973/work
pexpect==4.9.0
pickleshare==0.7.5
Pillow==10.1.0
pkgutil_resolve_name==1.3.10
platformdirs @ file:///croot/platformdirs_1692205439124/work
plotly==5.22.0
plyfile==1.0.2
pointnet2==0.0.0
pooch @ file:///croot/pooch_1695850093751/work
portalocker==2.8.2
prompt-toolkit==3.0.43
protobuf==3.20.3
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1 @ file:///Users/ktietz/demo/mc3/conda-bld/pyasn1_1629708007385/work
pyasn1-modules==0.2.8
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pyface==8.0.0
Pygments==2.17.2
PyJWT @ file:///opt/conda/conda-bld/pyjwt_1657544592787/work
pyOpenSSL @ file:///croot/pyopenssl_1690223430423/work
pyparsing==3.1.1
PyQt5==5.15.9
pyqt5-plugins==5.15.9.2.3
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
pyqt5-tools==5.15.9.3.3
pyquaternion==0.9.9
PySocks @ file:///tmp/build/80754af9/pysocks_1605305779399/work
python-dateutil==2.8.2
python-dotenv==1.0.1
python-utils==3.5.2
pytorch-msssim==1.0.0
pytorch3d @ git+https://github.com/facebookresearch/pytorch3d.git@47d5dc88247035b35ca3cfce159565f92d8fbb75
pytz==2023.3.post1
PyWavelets @ file:///home/conda/feedstock_root/build_artifacts/pywavelets_1695567543138/work
PyYAML @ file:///home/conda/feedstock_root/build_artifacts/pyyaml_1695373436676/work
qt5-applications==5.15.2.2.3
qt5-tools==5.15.2.1.3
ransac-voting-3d==0.0.0
referencing==0.35.1
requests @ file:///croot/requests_1690400202158/work
requests-oauthlib==1.3.0
retrying==1.3.4
rpds-py==0.18.1
rsa @ file:///tmp/build/80754af9/rsa_1614366226499/work
scikit-image @ file:///home/conda/feedstock_root/build_artifacts/scikit-image_1667117154325/work
scikit-learn==1.3.2
scipy==1.10.1
SimpleITK==2.3.1
six==1.15.0
smmap==5.0.1
stack-data==0.6.3
tabulate==0.9.0
tenacity==8.3.0
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit @ file:///home/builder/tkoch/workspace/tensorflow/tensorboard-plugin-wit_1658918494740/work/tensorboard_plugin_wit-1.8.1-py3-none-any.whl
tensorboardX @ file:///home/conda/feedstock_root/build_artifacts/tensorboardx_1692548726750/work
tensorflow @ file:///home/builder/ktietz/aggregate/tensorflow_recipes/ci_cpu/tensorflow-base_1614354478017/work/tensorflow_pkg/tensorflow-2.4.1-cp38-cp38-linux_x86_64.whl
tensorflow-estimator==2.4.0
termcolor==1.1.0
threadpoolctl==3.5.0
tifffile==2019.7.26.2
toolz @ file:///home/conda/feedstock_root/build_artifacts/toolz_1706112571092/work
torch==1.12.1+cu116
torch3d==0.1
torchaudio==0.12.1+cu116
torchgeometry==0.1.2
torchvision==0.13.1+cu116
tqdm==4.66.1
traitlets==5.14.3
traits==6.4.3
traitsui==8.0.0
typing_extensions==4.11.0
tzdata==2023.3
urllib3 @ file:///croot/urllib3_1698257533958/work
vtk==9.3.0
wcwidth==0.2.13
Werkzeug==3.0.3
widgetsnbextension==4.0.10
wrapt==1.12.1
yacs==0.1.8
yarl @ file:///croot/yarl_1701105127787/work
zipp @ file:///croot/zipp_1672387121353/work
